% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.cimg.array}
\alias{as.cimg.array}
\title{Turn an numeric array into a cimg object}
\usage{
\method{as.cimg}{array}(obj, ...)
}
\arguments{
\item{obj}{an array}

\item{...}{ignored}
}
\description{
If the array has two dimensions, we assume it's a grayscale image. If it has three dimensions we assume it's a video, unless the third dimension has a depth of 3, in which case we assume it's a colour image,
}
\examples{
as.cimg(array(1:9,c(3,3)))
as.cimg(array(1,c(10,10,3))) #Guesses colour image
as.cimg(array(1:9,c(10,10,4))) #Guesses video
}
