% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.cimg.im}
\alias{as.cimg.im}
\title{Convert an image in spatstat format to an image in cimg format}
\usage{
\method{as.cimg}{im}(obj, ...)
}
\arguments{
\item{obj}{a spatstat image}

\item{...}{optional arguments}
}
\value{
a cimg image
}
\description{
Convert an image in spatstat format to an image in cimg format
}
\author{
Simon Barthelme
}
