% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{as.cimg}
\alias{as.cimg}
\alias{as.cimg.matrix}
\title{Convert to cimg object}
\usage{
as.cimg(obj, ...)

\method{as.cimg}{matrix}(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{optional arguments}
}
\description{
Imager implements various converters that turn your data into cimg objects
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: 
}}
\seealso{
as.cimg.array, as.cimg.function, as.cimg.data.frame
}

