% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{as.data.frame.imlist}
\alias{as.data.frame.imlist}
\title{Convert image list to data.frame}
\usage{
\method{as.data.frame}{imlist}(x, ..., index = "im")
}
\arguments{
\item{x}{an image list (an imlist object)}

\item{...}{Passed on to as.data.frame.cimg}

\item{index}{Name of the column containing the index (or name) of the image in the list. Default: "im"}
}
\description{
Convert image list to data.frame
}
\examples{
#Transform the image gradient into a data.frame
gr <- imgradient(boats,"xy") \%>\% setNames(c("dx","dy")) \%>\% as.data.frame
str(gr)
}
