% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{as.data.frame.pixset}
\alias{as.data.frame.pixset}
\alias{convert_pixset}
\title{Methods to convert pixsets to various objects}
\usage{
\method{as.data.frame}{pixset}(x, ..., drop = FALSE)
}
\arguments{
\item{x}{pixset to convert}

\item{...}{ignored}

\item{drop}{drop flat dimensions}
}
\description{
Methods to convert pixsets to various objects
}
\examples{

px <- boats > 250
#Convert to array of logicals
as.logical(px) \%>\% dim
#Convert to data.frame: gives all pixel locations in the set
as.data.frame(px) \%>\% head
#Drop flat dimensions
as.data.frame(px,drop=TRUE) \%>\% head
}
\seealso{
where
}
