% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{as.imlist.list}
\alias{as.imlist}
\alias{as.imlist.cimg}
\alias{as.imlist.imlist}
\alias{as.imlist.list}
\alias{as.list.imlist}
\title{Convert various objects to image lists}
\usage{
\method{as.imlist}{list}(l)

as.imlist(x, ...)

\method{as.list}{imlist}(x, ...)

\method{as.imlist}{imlist}(x, ...)

\method{as.imlist}{cimg}(x, ...)
}
\arguments{
\item{x}{an image list}

\item{...}{ignored}
}
\value{
a list
}
\description{
Convert various objects to image lists
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: convert from list

\item \code{imlist}: Convert from list

\item \code{imlist}: Convert from imlist (identity)

\item \code{cimg}: Convert from image
}}
\examples{
list(a=boats,b=boats*2) \%>\% as.imlist
}

