% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{as.list.imlist}
\alias{as.list.imlist}
\title{Convert image list to list}
\usage{
\method{as.list}{imlist}(x, ...)
}
\arguments{
\item{x}{an image list}

\item{...}{ignored}
}
\value{
a list
}
\description{
Convert image list to list
}

