% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{as.pixset}
\alias{as.cimg.pixset}
\alias{as.pixset}
\alias{as.pixset.cimg}
\title{Methods to convert various objects to pixsets}
\usage{
as.pixset(x, ...)

\method{as.pixset}{cimg}(x, ...)

\method{as.cimg}{pixset}(x, ...)
}
\arguments{
\item{x}{object to convert to pixset}

\item{...}{ignored}
}
\description{
Methods to convert various objects to pixsets
}
\section{Methods (by class)}{
\itemize{
\item \code{cimg}: convert cimg to pixset

\item \code{pixset}: convert pixset to pixset (nothing happens)
}}
\examples{
#When converting an image to a pixset, the default is to include all pixels with non-zero value 
as.pixset(boats)
#The above is equivalent to:
boats!=0

}

