% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.raster.cimg}
\alias{as.raster.cimg}
\title{Convert a cimg object to a raster object}
\usage{
\method{as.raster}{cimg}(x, frames, rescale.color = TRUE, ...)
}
\arguments{
\item{x}{an image (of class cimg)}

\item{frames}{which frames to extract (in case depth > 1)}

\item{rescale.color}{rescale so that pixel values are in [0,1]? (subtract min and divide by range). default TRUE}

\item{...}{ignored}
}
\value{
a raster object
}
\description{
raster objects are used by R's base graphics for plotting
}
\author{
Simon Barthelme
}
\seealso{
plot.cimg, rasterImage
}

