% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{boats}
\alias{boats}
\title{Photograph of sailing boats from Kodak set}
\format{
an image of class cimg
}
\source{
\url{ http://r0k.us/graphics/kodak/kodim09.html }
}
\usage{
boats
}
\description{
This photograph was downloaded from http://r0k.us/graphics/kodak/kodim09.html. Its size was reduced by half to speed up loading and save space.
}
\keyword{datasets}
