% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{boundary}
\alias{boundary}
\title{Find the boundary of a shape in a pixel set}
\usage{
boundary(px)
}
\arguments{
\item{px}{pixel set}
}
\description{
Find the boundary of a shape in a pixel set
}
\examples{
px.diamond(10,30,30) \%>\% boundary \%>\% plot
px.square(10,30,30) \%>\% boundary \%>\% plot
}

