% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boxblur}
\alias{boxblur}
\title{Blur image with a box filter (square window)}
\usage{
boxblur(im, sigma, boundary_conditions = TRUE)
}
\arguments{
\item{im}{an image}

\item{sigma}{Size of the box window.}

\item{boundary_conditions}{Boundary conditions. FALSE: Dirichlet TRUE: Neumann.}
}
\description{
Blur image with a box filter (square window)
}
\examples{
boxblur(boats,5) \%>\% plot(main="Dirichlet boundary")
boxblur(boats,5,TRUE) \%>\% plot(main="Neumann boundary")
}
\seealso{
deriche(), vanvliet().
}

