% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boxblur_xy}
\alias{boxblur_xy}
\title{Blur image with a box filter.}
\usage{
boxblur_xy(im, sx, sy, boundary_conditions = TRUE)
}
\arguments{
\item{im}{an image
@param sx Size of the box window, along the X-axis.
@param sy Size of the box window, along the Y-axis.
@param boundary_conditions Boundary conditions. Can be <tt>{ false=dirichlet | true=neumann }</tt>.
@seealso blur().}
}
\description{
This is a recursive algorithm, not depending on the values of the box kernel size.
}
\examples{
boxblur_xy(boats,20,5) \%>\% plot(main="Anisotropic blur")
}

