% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bucket_fill}
\alias{bucket_fill}
\title{Bucket fill}
\usage{
bucket_fill(im, x, y, z, color, opacity = 1, sigma = 0,
  is_high_connexity = FALSE)
}
\arguments{
\item{x}{X-coordinate of the starting point of the region to fill.}

\item{y}{Y-coordinate of the starting point of the region to fill.}

\item{z}{Z-coordinate of the starting point of the region to fill.}

\item{color}{Pointer to spectrum() consecutive values, defining the drawing color.}

\item{opacity}{Opacity of the drawing.}

\item{sigma}{Tolerance concerning neighborhood values.}

\item{is_high_connexity}{Use 8-connexity (only for 2d images).}
}
\description{
Bucket fill
}

