% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bucket_select}
\alias{bucket_select}
\title{Select a region of homogeneous colour}
\usage{
bucket_select(im, x, y, z, sigma = 0, is_high_connexity = FALSE)
}
\arguments{
\item{x}{X-coordinate of the starting point of the region to fill.}

\item{y}{Y-coordinate of the starting point of the region to fill.}

\item{z}{Z-coordinate of the starting point of the region to fill.}

\item{sigma}{Tolerance concerning neighborhood values.}

\item{is_high_connexity}{Use 8-connexity (only for 2d images).}
}
\description{
The underlying algorithm is the same as the bucket fill (AKA flood fill). Unlike with the bucket fill, the image isn't changed, the function simply returns a binary mask of the selected region
}

