% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{capture.plot}
\alias{capture.plot}
\title{Capture the current R plot device as a cimg image}
\usage{
capture.plot()
}
\value{
a cimg image corresponding to the contents of the current plotting window
}
\description{
Capture the current R plot device as a cimg image
}
\examples{
plot(1:10)
capture.plot() \%>\% plot #A plot of the plot
}
\author{
Simon Barthelme
}

