% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stencils.R
\name{center.stencil}
\alias{center.stencil}
\title{Center stencil at a location}
\usage{
center.stencil(stencil, ...)
}
\arguments{
\item{stencil}{a stencil (data.frame with coordinates dx,dy,dz,dc)}

\item{...}{centering locations (e.g. x=4,y=2)}
}
\description{
Center stencil at a location
}
\examples{
stencil <- data.frame(dx=seq(-2,2,1),dy=seq(-2,2,1))
center.stencil(stencil,x=10,y=20)
}
