% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{cimg.dimensions}
\alias{cimg.dimensions}
\alias{width}
\alias{height}
\alias{spectrum}
\alias{depth}
\alias{nPix}
\title{Image dimensions}
\usage{
width(im)

height(im)

spectrum(im)

depth(im)

nPix(im)
}
\arguments{
\item{im}{an image}
}
\description{
Image dimensions
}
\section{Functions}{
\itemize{
\item \code{width}: Width of the image (in pixels)

\item \code{height}: Height of the image (in pixels)

\item \code{spectrum}: Number of colour channels

\item \code{depth}: Depth of the image/number of frames in a video

\item \code{nPix}: Total number of pixels (prod(dim(im)))
}}

