% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{cimg.extract}
\alias{B}
\alias{G}
\alias{R}
\alias{channel}
\alias{cimg.extract}
\title{Various shortcuts for extracting colour channels, frames, etc}
\usage{
channel(im, ind)

R(im)

G(im)

B(im)
}
\arguments{
\item{im}{an image}

\item{ind}{channel index}

\item{index}{frame index}
}
\description{
Various shortcuts for extracting colour channels, frames, etc
}
\section{Functions}{
\itemize{
\item \code{channel}: Extract an image channel

\item \code{R}: Extract red channel

\item \code{G}: Extract green channel

\item \code{B}: Extract blue channel
}}

