% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cimg.use.openmp}
\alias{cimg.use.openmp}
\title{Control CImg's parallelisation}
\usage{
cimg.use.openmp(mode = "adaptive", nthreads = 1)
}
\arguments{
\item{mode}{Either "adaptive","always" or "none". The default is adaptive (parallelisation for large images only).}

\item{nthreads}{The number of threads to use when running code in parallel. The default is 1.}
}
\value{
NULL (function is used for side effects)
}
\description{
On supported architectures CImg can parallelise many operations using OpenMP (e.g. \code{\link{imager.combine}}). Use this function to turn parallelisation on or off.
}
\details{
You need to be careful that \option{nthreads} is not higher than the value in the system environment variable OMP_THREAD_LIMIT (this can be checked with Sys.getenv('OMP_THREAD_LIMIT')). The OMP_THREAD_LIMIT thread limit usually needs to be correctly set before launching R, so using Sys.setenv once a session has started is not certain to work.
}
\examples{
cimg.use.openmp("never") #turn off parallelisation
}
\author{
Simon Barthelme
}
