% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hough.R
\name{circles}
\alias{circles}
\title{Add circles to plot}
\usage{
circles(x, y, radius, bg = NULL, fg = "white", ...)
}
\arguments{
\item{x}{centers (x coordinate)}

\item{y}{centers (y coordinate)}

\item{radius}{radius (in user coordinates)}

\item{bg}{background colour}

\item{fg}{foreground colour}

\item{...}{passed to polygon, e.g. lwd}
}
\value{
none, used for side effect
}
\description{
Base R has a function for plotting circles called "symbols". Unfortunately, the size of the circles is inconsistent across devices. This function plots circles whose radius is specified in used coordinates.
}
\seealso{
hough_circle
}
\author{
Simon Barthelme
}
