% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diffusion_tensors}
\alias{diffusion_tensors}
\title{Compute field of diffusion tensors for edge-preserving smoothing.}
\usage{
diffusion_tensors(im, sharpness = 0.7, anisotropy = 0.6, alpha = 0.6,
  sigma = 1.1, is_sqrt = FALSE)
}
\description{
@param im an image
      @param sharpness Sharpness
      @param anisotropy Anisotropy
      @param alpha Standard deviation of the gradient blur.
      @param sigma Standard deviation of the structure tensor blur.
      @param is_sqrt Tells if the square root of the tensor field is computed instead.
}

