% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dilate}
\alias{dilate}
\title{Dilate image by a structuring element.}
\usage{
dilate(im, mask, boundary_conditions = TRUE, normalise = FALSE)
}
\arguments{
\item{im}{an image}

\item{mask}{Structuring element.}

\item{boundary_conditions}{Boundary conditions.}

\item{normalise}{Normalise mask (default FALSE)}
}
\description{
Dilate image by a structuring element.
}

