% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dilate_rect}
\alias{dilate_rect}
\title{Dilate image by a rectangular structuring element of specified size.}
\usage{
dilate_rect(im, sx, sy, sz = 1L)
}
\arguments{
\item{sx}{Width of the structuring element.}

\item{sy}{Height of the structuring element.}

\item{sz}{Depth of the structuring element.}
}
\description{
Dilate image by a rectangular structuring element of specified size.
}

