% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dilate_square}
\alias{dilate_square}
\title{Dilate image by a square structuring element of specified size.}
\usage{
dilate_square(im, size)
}
\arguments{
\item{size}{Size of the structuring element.}
}
\description{
Dilate image by a square structuring element of specified size.
}

