% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{display.cimg}
\alias{display.cimg}
\title{Display image using CImg library}
\usage{
\method{display}{cimg}(x, ..., rescale = TRUE)
}
\arguments{
\item{x}{an image (cimg object)}

\item{...}{ignored}

\item{rescale}{if true pixel values are rescaled to [0-1] (default TRUE)}
}
\description{
Press escape or close the window to exit.
}
\examples{
##Not run: interactive only 
##display(boats,TRUE) #Normalisation on 
##display(boats/2,TRUE) #Normalisation on, so same as above
##display(boats,FALSE) #Normalisation off
##display(boats/2,FALSE) #Normalisation off, so different from above
}
