% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{display.list}
\alias{display.list}
\title{Display image list using CImg library}
\usage{
\method{display}{list}(im, rescale = TRUE)
}
\arguments{
\item{imlist}{a list of cimg objects}
}
\description{
Display image list using CImg library
}

