% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erode}
\alias{erode}
\title{Erode image by a structuring element.}
\usage{
erode(im, mask, boundary_conditions = TRUE, normalise = FALSE)
}
\arguments{
\item{im}{an image}

\item{mask}{Structuring element.}

\item{boundary_conditions}{Boundary conditions.}

\item{normalise}{Determines if the closing is locally normalised (default FALSE)}
}
\description{
Erode image by a structuring element.
}

