% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erode_square}
\alias{erode_square}
\title{Erode image by a  square structuring element of specified size.}
\usage{
erode_square(im, size)
}
\arguments{
\item{im}{an image}

\item{size}{size of the structuring element.}
}
\description{
Erode image by a  square structuring element of specified size.
}

