% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_patches_min}
\alias{extract_patches_min}
\title{Return minimum of image patches}
\usage{
extract_patches_min(im, cx, cy, wx, wy)
}
\arguments{
\item{im}{an image}

\item{cx}{vector of x coordinates for patch centers}

\item{cy}{vector of y coordinates for patch centers}

\item{wx}{vector of coordinates for patch width}

\item{wy}{vector of coordinates for patch height}
}
\value{
a numeric vector of the minimum values of the patches
}
\description{
Patches are rectangular (cubic) image regions centered at cx,cy (cz) with width wx and height wy (opt. depth wz)
}

