% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{frame}
\alias{frame}
\title{Extract one frame out of a 4D image/video}
\usage{
frame(im, index)
}
\arguments{
\item{im}{an image}

\item{index}{frame index}
}
\value{
an image (class cimg)
}
\description{
Extract one frame out of a 4D image/video
}
\author{
Simon Barthelme
}

