% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{frames}
\alias{frames}
\title{Split a video into separate frames}
\usage{
frames(im, index, drop = FALSE)
}
\arguments{
\item{im}{an image}

\item{index}{which channels to extract (default all)}

\item{drop}{if TRUE drop extra dimensions, returning normal arrays and not cimg objects}
}
\value{
a list of frames
}
\description{
Split a video into separate frames
}
\seealso{
channels
}

