% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_gradient}
\alias{get_gradient}
\title{Compute image gradient.}
\usage{
get_gradient(im, axes = "", scheme = 3L)
}
\arguments{
\item{im}{an image}

\item{axes}{Axes considered for the gradient computation, as a C-string (e.g "xy").}

\item{scheme}{= Numerical scheme used for the gradient computation:
1 = Backward finite differences
0 = Centered finite differences
1 = Forward finite differences
2 = Using Sobel masks
3 = Using rotation invariant masks
4 = Using Deriche recursive filter.
5 = Using Van Vliet recursive filter.}
}
\value{
a list of images (corresponding to the different directions)
}
\description{
Compute image gradient.
}

