% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_hessian}
\alias{get_hessian}
\title{Return image hessian.}
\usage{
get_hessian(im, axes = "")
}
\arguments{
\item{im}{an image}

\item{axes}{Axes considered for the hessian computation, as a character string (e.g "xy").}
}
\description{
Return image hessian.
}

