% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{grab}
\alias{grab}
\alias{grabLine}
\alias{grabPoint}
\alias{grabRect}
\title{Select image regions interactively}
\usage{
grabLine(im, coord = TRUE)

grabRect(im, coord = TRUE)

grabPoint(im, coord = TRUE)
}
\arguments{
\item{im}{an image}

\item{coord}{if TRUE, return coordinates. if FALSE, content}
}
\value{
either a vector of coordinates, or an image
}
\description{
These functions let you select a shape in an image (a point, a line, or a rectangle)
They either return the coordinates of the shape (default), or the contents.
In case of lines contents are interpolated.
}
\examples{
##Not run: interactive only 
##grabRect(boats)
##grabRect(boats,TRUE)
}
\author{
Simon Barthelme
}
\seealso{
display
}

