% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%inr\%}
\alias{\%inr\%}
\title{Check that value is in a range}
\usage{
x \%inr\% range
}
\arguments{
\item{x}{numeric values}

\item{range}{a vector of length two, of the form c(a,b)}
}
\value{
a vector of logicals
1:10 %inr% c(0,5)
}
\description{
A shortcut for x >= a | x <= b.
}
\author{
Simon Barthelme
}
