% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{gsdim}
\alias{gsdim}
\title{Grayscale dimensions of image}
\usage{
gsdim(im)
}
\arguments{
\item{im}{an image}
}
\value{
returns c(dim(im)[1:3],1)
}
\description{
Shortcut, returns the dimensions of an image if it had only one colour channel
}
\examples{
imnoise(dim=gsdim(boats))
}
\author{
Simon Barthelme
}
