% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitcombine.R
\name{idply}
\alias{idply}
\title{Split an image along axis, map function, return a data.frame}
\usage{
idply(im, axis, fun, ...)
}
\arguments{
\item{im}{image}

\item{axis}{axis for the split (e.g "c")}

\item{fun}{function to apply}

\item{...}{extra arguments to function fun}
}
\description{
Shorthand for imsplit followed by purrr::map_df
}
\examples{
idply(boats,"c",mean) #mean luminance per colour channel
}
