% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitcombine.R
\name{ilply}
\alias{ilply}
\title{Split an image along axis, apply function, return a list}
\usage{
ilply(im, axis, fun, ...)
}
\arguments{
\item{im}{image}

\item{axis}{axis for the split (e.g "c")}

\item{fun}{function to apply}

\item{...}{extra arguments for function fun}
}
\description{
Shorthand for imsplit followed by purrr::map
}
\examples{
parrots <- load.example("parrots")
ilply(parrots,"c",mean) #mean luminance per colour channel
}
