% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{im_split}
\alias{im_split}
\title{Split an image along a certain axis (producing a list)}
\usage{
im_split(im, axis, nb = -1L)
}
\arguments{
\item{im}{an image}

\item{axis}{the axis along which to split (for example 'c')}

\item{nb}{number of objects to split into.
if nb=-1 (the default) the maximum number of splits is used ie. split(im,"c") produces a list containing all individual colour channels}
}
\description{
Split an image along a certain axis (producing a list)
}
\seealso{
imappend (the reverse operation)
}

