% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cimg_class.R
\name{RGBtoHSL}
\alias{HSItoRGB}
\alias{HSLtoRGB}
\alias{HSVtoRGB}
\alias{RGBtoHSI}
\alias{RGBtoHSL}
\alias{RGBtoHSV}
\alias{RGBtoYCbCr}
\alias{RGBtoYUV}
\alias{RGBtosRGB}
\alias{YCbCrtoRGB}
\alias{YUVtoRGB}
\alias{imager.colourspaces}
\alias{sRGBtoRGB}
\title{Colour space conversions in imager}
\usage{
RGBtoHSL(im)

HSLtoRGB(im)

RGBtoHSV(im)

HSVtoRGB(im)

RGBtoHSI(im)

HSItoRGB(im)

RGBtosRGB(im)

sRGBtoRGB(im)

RGBtoYCbCr(im)

YCbCrtoRGB(im)

RGBtoYUV(im)

YUVtoRGB(im)
}
\arguments{
\item{im}{an image}
}
\description{
All functions listed here assume the input image has three colour channels (spectrum(im) == 3)
}
\section{Functions}{
\itemize{
\item \code{RGBtoHSL}: RGB to HSL conversion

\item \code{HSLtoRGB}: HSL to RGB conversion

\item \code{RGBtoHSV}: RGB to HSV conversion

\item \code{HSVtoRGB}: HSV to RGB conversion

\item \code{RGBtoHSI}: RGB to HSI conversion

\item \code{HSItoRGB}: HSI to RGB conversion

\item \code{RGBtosRGB}: RGB to sRGB conversion

\item \code{sRGBtoRGB}: sRGB to RGB conversion

\item \code{RGBtoYCbCr}: RGB to YCbCr conversion

\item \code{YCbCrtoRGB}: YCbCr to RGB conversion

\item \code{RGBtoYUV}: RGB to YUV conversion

\item \code{YUVtoRGB}: YUV to RGB conversion
}}

