% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imageR.R
\name{imageROutput}
\alias{imageROutput}
\title{Widget output function for use in Shiny}
\usage{
imageROutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

