% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imappend}
\alias{imappend}
\title{Combine a list of images into a single image}
\usage{
imappend(imlist, axis)
}
\arguments{
\item{imlist}{a list of images (all elements must be of class cimg)}

\item{axis}{the axis along which to split (for example 'c')}
}
\description{
All images will be concatenated along the x,y,z, or c axis.
}
\seealso{
imsplit (the reverse operation)
}

