% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imdraw}
\alias{imdraw}
\title{Draw image on another image}
\usage{
imdraw(im, sprite, x = 1, y = 1, z = 1, opacity = 1)
}
\arguments{
\item{im}{background image}

\item{sprite}{sprite to draw on background image}

\item{x}{location}

\item{y}{location}

\item{z}{location}

\item{opacity}{transparency level (default 1)}
}
\description{
Draw image on another image
}
\examples{
im <- load.example("parrots")
boats.small <- imresize(boats,.5)
#I'm aware the result is somewhat ugly
imdraw(im,boats.small,x=400,y=10,opacity=.7) \%>\% plot
}
\seealso{
imager.combine, for different ways of combining images
}
\author{
Simon Barthelme
}
