% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imgradient}
\alias{imgradient}
\title{Compute image gradient}
\usage{
imgradient(im, axes = "xy", scheme = 3)
}
\arguments{
\item{im}{an image of class cimg}

\item{axes}{direction along which to compute the gradient. Either a single character (e.g. "x"), or multiple characters (e.g. "xyz"). Default: "xy"}

\item{scheme}{numerical scheme (default '3', rotation invariant)}
}
\value{
an image or a list of images, depending on the value of "axes"
}
\description{
Light interface for get_gradient. Refer to get_gradient for details on the computation.
}
\examples{
grayscale(boats) \%>\% imgradient("x") \%>\% plot
imgradient(boats,"xy") #Returns a list
}
\author{
Simon Barthelme
}
