% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imhessian}
\alias{imhessian}
\title{Compute image hessian.}
\usage{
imhessian(im, axes = c("xx", "xy", "yy"))
}
\arguments{
\item{im}{an image}

\item{axes}{Axes considered for the hessian computation, as a character string (e.g "xy" corresponds to d/(dx*dy)). Can be a list of axes. Default: xx,xy,yy}
}
\value{
an image, or a list of images
}
\description{
Compute image hessian.
}
\examples{
imhessian(boats,"xy") \%>\% plot(main="Second-derivative, d/(dx*dy)")
}
