% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imlap}
\alias{imlap}
\title{Compute image Laplacian}
\usage{
imlap(im)
}
\arguments{
\item{im}{an image}
}
\description{
The Laplacian is the sum of second derivatives, approximated here using finite differences.
}
\examples{
imlap(boats) \%>\% plot
}
