% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{imrep}
\alias{imrep}
\title{Replicate images}
\usage{
imrep(x, n = 1, axis = NULL)
}
\arguments{
\item{x}{an image}

\item{n}{number of replications}

\item{axis}{axis to append along (one of NULL, "x","y","z","c"). Default: NULL}
}
\value{
either an image or an image list
}
\description{
Kinda like rep, for images. Copy image n times and (optionally), append.
}
\examples{
#Result is a list
imrep(boats,3) \%>\% plot
#Result is an image 
imrep(boats,3,"x") \%>\% plot
#Make an animation by repeating each frame 10x
#map_il(1:5,~ isoblur(boats,.) \%>\% imrep(10,"z")) \%>\%
#                       imappend("z") \%>\% play
}
\author{
Simon Barthelme
}
