% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imsharpen}
\alias{imsharpen}
\title{Sharpen image.}
\usage{
imsharpen(im, amplitude, sharpen_type = FALSE, edge = 1, alpha = 0,
  sigma = 0)
}
\description{
@param im an image
      @param amplitude Sharpening amplitude
      @param sharpen_type Select sharpening method. Can be <tt>{ false=inverse diffusion | true=shock filters }</tt>.
      @param edge Edge threshold (shock filters only).
      @param alpha Gradient smoothness (shock filters only).
      @param sigma Tensor smoothness (shock filters only).
}
\examples{
layout(t(1:2))
plot(boats,main="Original")
imsharpen(boats,150)  \%>\% plot(main="Sharpened")

}

