% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imshift}
\alias{imshift}
\title{Shift image content.}
\usage{
imshift(im, delta_x = 0L, delta_y = 0L, delta_z = 0L, delta_c = 0L,
  boundary_conditions = 0L)
}
\arguments{
\item{im}{an image
@param delta_x Amount of displacement along the X-axis.
@param delta_y Amount of displacement along the Y-axis.
@param delta_z Amount of displacement along the Z-axis.
@param delta_c Amount of displacement along the C-axis.
@param boundary_conditions can be:
   - 0: Zero border condition (Dirichlet).
   - 1: Nearest neighbors (Neumann).
   - 2: Repeat Pattern (Fourier style).}
}
\description{
Shift image content.
}
\examples{
imshift(boats,10,50) \%>\% plot
}

