% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imageR.R
\name{intenseOutput}
\alias{intenseOutput}
\title{Widget output function for use in Shiny}
\usage{
intenseOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

