% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{is.imlist}
\alias{is.imlist}
\title{Check that an object is an imlist object}
\usage{
is.imlist(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical
}
\description{
Check that an object is an imlist object
}
