% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{is.pixset}
\alias{is.pixset}
\title{Check that an object is a pixset object}
\usage{
is.pixset(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical
}
\description{
Check that an object is a pixset object
}
