% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isoblur}
\alias{isoblur}
\title{Blur image isotropically.}
\usage{
isoblur(im, sigma, boundary_conditions = TRUE, gaussian = FALSE)
}
\arguments{
\item{im}{an image}

\item{sigma}{Standard deviation of the blur.}

\item{boundary_conditions}{Boundary conditions. Can be <tt>{ 0=dirichlet | 1=neumann }}

\item{gaussian}{}
}
\description{
Blur image isotropically.
}
\seealso{
deriche(), vanvliet().
}

