% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitcombine.R
\name{liply}
\alias{liply}
\title{Apply function to each element of a list, then combine the result as an image by appending along specified axis}
\usage{
liply(lst, fun, axis, ...)
}
\arguments{
\item{lst}{a list}

\item{fun}{function to apply}

\item{axis}{which axis to append along (e.g. "c" for colour)}

\item{...}{further arguments to be passed to fun}
}
\description{
This is just a shortcut for purrr::map followed by imappend
}
\examples{
build.im <- function(size) as.cimg(function(x,y) (x+y)/size,size,size)
liply(c(10,50,100),build.im,"y") \%>\% plot
}
