% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{load.image}
\alias{load.image}
\title{Load image from file}
\usage{
load.image(file)
}
\arguments{
\item{file}{path to file}
}
\value{
an object of class 'cimg'
}
\description{
You'll need ImageMagick for some formats.
}

