% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{magick}
\alias{as.cimg.magick-image}
\alias{as.imlist.magick-image}
\alias{magick}
\title{Convert a magick image to a cimg image or image list}
\usage{
\method{as.imlist}{magick-image}(obj, alpha = "rm", ...)

\method{as.cimg}{magick-image}(obj, alpha = "rm", ...)
}
\arguments{
\item{obj}{an object of class "magick-image"}

\item{alpha}{what do to with the alpha channel ("rm": remove and store as attribute, "flatten": flatten, "keep": keep). Default: "rm"}

\item{...}{ignored}
}
\value{
an object of class cimg or imlist
}
\description{
The magick library package stores its data as "magick-image" object, which may in fact contain several images or an animation. These functions convert magick objects into imager objects.
}
\author{
Jan Wijffels, Simon Barthelme
}
\seealso{
flatten.alpha, rm.alpha
}

