% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mclosing}
\alias{mclosing}
\title{Morphological closing (dilation followed by erosion)}
\usage{
mclosing(im, mask, boundary_conditions = TRUE, normalise = FALSE)
}
\arguments{
\item{im}{an image}

\item{mask}{Structuring element.}

\item{boundary_conditions}{Boundary conditions.}

\item{normalise}{Determines if the closing is locally normalised (default FALSE)}
}
\description{
Morphological closing (dilation followed by erosion)
}

