% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mclosing_square}
\alias{mclosing_square}
\title{Morphological closing by a square element (dilation followed by erosion)}
\usage{
mclosing_square(im, size)
}
\arguments{
\item{size}{size of the square element}
}
\description{
Morphological closing by a square element (dilation followed by erosion)
}

