% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mirror}
\alias{mirror}
\title{Mirror image content along specified axis}
\usage{
mirror(im, axis)
}
\arguments{
\item{im}{an image}

\item{axis}{Mirror axis ("x","y","z","c")}
}
\description{
Mirror image content along specified axis
}
\examples{
mirror(boats,"x") \%>\% plot
mirror(boats,"y") \%>\% plot
}
