% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mopening}
\alias{mopening}
\title{Morphological opening (erosion followed by dilation)}
\usage{
mopening(im, mask, boundary_conditions = TRUE, normalise = FALSE)
}
\arguments{
\item{im}{an image}

\item{mask}{Structuring element.}

\item{boundary_conditions}{Boundary conditions.}

\item{normalise}{Determines if the closing is locally normalised (default FALSE)}
}
\description{
Morphological opening (erosion followed by dilation)
}

