% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mopening_square}
\alias{mopening_square}
\title{Morphological opening by a square element (erosion followed by dilation)}
\usage{
mopening_square(im, size)
}
\arguments{
\item{im}{an image}

\item{size}{size of the square element}
}
\description{
Morphological opening by a square element (erosion followed by dilation)
}

