% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nlmeans}
\alias{nlmeans}
\title{Denoising by non-local means}
\usage{
nlmeans(im, patch_size = 1L, lambda = -1, sigma = -1, alpha = 3L,
  sampling = 1L)
}
\arguments{
\item{patch_size}{Patch size (in pixels, default: 1)}

\item{lambda}{Bandwidth (default: -1, auto)}

\item{sigma}{Noise standard deviation (default: -1, auto)}

\item{sampling}{Ignore some proportion of the pixels in the patch, e.g. sampling = 2 uses only every other pixel, sampling = 3 every third, etc.}
}
\description{
Performs denoising by averaging over image neighbourhoods defined by patch similarity. The algorithm can work very well but it is extremely slow for large patch size. You should set the "sampling" parameter to a high value if you're going to use large patches. 
This function wraps a CImg plugin by Jerome Boulanger.
}
\examples{
layout(t(1:2))
bts <- grayscale(boats)
bts.noisy <- bts+imnoise(dim=dim(bts))*.1
plot(bts.noisy,main="Original")
nlmeans(bts.noisy,patch=1) \%>\% plot(main="Denoising by nlmeans")
}

