% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{patch_summary}
\alias{patch_summary}
\title{Extract a numerical summary from image patches}
\usage{
patch_summary(im, expr, cx, cy, wx, wy)
}
\description{
Extract a numerical summary from image patches
}
\examples{
#Example: median filtering using patch_summary
#Center a patch at each pixel
im <- grayscale(boats)
patches <- pixel.grid(im)  \%>\% mutate(w=3,h=3)
#Extract patch summary:
out <- mutate(patches,med=patch_summary(im,"ic",x,y,w,h))
as.cimg(out,v.name="med") \%>\% plot
}

