% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_axes}
\alias{permute_axes}
\title{Permute image axes}
\usage{
permute_axes(im, perm)
}
\arguments{
\item{im}{an image}

\item{perm}{a character string, e.g., "zxyc" to have the z-axis come first}
}
\description{
By default images are stored in xyzc order. Use permute_axes to change that order.
}
\examples{
im <- array(0,c(10,30,40,3)) \%>\% as.cimg
permute_axes(im,"zxyc")
}
