% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{pixel.grid}
\alias{pixel.grid}
\title{Returns the pixel grid for an image}
\usage{
pixel.grid(im)
}
\arguments{
\item{im}{}
}
\value{
a data.frame
}
\description{
The pixel grid for image im gives the (x,y,z,c) coordinates of each successive pixel as a data.frame. The c coordinate has been renamed 'cc' to avoid conflicts with R's c function.
NB: coordinates start at (x=1,y=1), corresponding to the top left corner of the image
}

