% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{plot.imlist}
\alias{plot.imlist}
\title{Plot an image list}
\usage{
\method{plot}{imlist}(x, layout = "rect", ...)
}
\arguments{
\item{x}{an image list (of type imlist)}

\item{layout}{either a matrix (in the format defined by the layout command) or one of "row","col" or "rect". Default: "rect"}

\item{...}{other parameters, to be passed to the plot command}
}
\description{
Each image in the list will be plotted separately. The layout argument controls the overall layout of the plot window.
}
\examples{
imsplit(boats,"c") #Returns an image list

}
\author{
Simon Barthelme
}

