% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pmedian}
\alias{pmedian}
\title{Compute the pixel-wise median across an image list}
\usage{
pmedian(x)
}
\arguments{
\item{x}{an image list}
}
\description{
Similarly to other reductions like "average", "parmax", and "parmin", this function computes the pixel-wise median in an image list (i.e., the median value for each pixel when comparing pixels across the list).
Images of different sizes are supported, in which case all images will be aligned to the first image, and padded with zeros.
}
\examples{
#Compute median values across colour channels
imsplit(boats,"c") \%>\% pmedian \%>\% plot
}

