% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{px.circle}
\alias{px.circle}
\title{Various useful pixsets}
\usage{
px.circle(r, x = 2 * r + 1, y = 2 * r + 1)
}
\arguments{
\item{r}{radius (in pixels)}

\item{x}{width (default 2*r+1)}

\item{y}{height (default 2*r+1)}
}
\value{
a pixset
px.circle(20,350,350) %>% plot(interp=FALSE)
px.circle(3) %>% plot(interp=FALSE)
r <- 5
layout(t(1:3))
plot(px.circle(r,20,20))
plot(px.square(r,20,20))
plot(px.diamond(r,20,20))
#They're useful as structuring elements
px <- grayscale(boats) > .8
grow(px,px.circle(5)) %>% plot
}
\description{
px.circle returns an (approximately) circular pixset of radius r, embedded in an image of width x and height y
Mathematically speaking, the set of all pixels whose L2 distance to the center equals r or less.
px.diamond is similar but returns a diamong (L1 distance less than r)
px.square is also similar but returns a square (Linf distance less than r)
}
\author{
Simon Barthelme
}

