% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{px.na}
\alias{px.na}
\title{A pixset for NA values}
\usage{
px.na(im)
}
\arguments{
\item{im}{an image}
}
\value{
a pixset
}
\description{
A pixset containing all NA pixels
}
\examples{
im <- boats
im[1] <- NA
px.na(im)
}
