% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{px.remove_outer}
\alias{px.remove_outer}
\title{Remove all connected regions that touch image boundaries}
\usage{
px.remove_outer(px)
}
\arguments{
\item{px}{a pixset}
}
\value{
a pixset
}
\description{
All pixels that belong to a connected region in contact with image boundaries are set to FALSE.
}
\examples{
im <- draw_circle(imfill(100,100),c(0,50,100),c(50,50,50),radius=10,color=1)
plot(im)
as.pixset(im) \%>\% px.remove_outer \%>\% plot
}
\author{
Simon Barthelme
}
