% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{RasterPackage}
\alias{RasterPackage}
\alias{as.cimg.RasterLayer}
\alias{as.imlist.RasterStackBrick}
\title{Convert a RasterLayer/RasterBrick to a cimg image/image list}
\usage{
\method{as.cimg}{RasterLayer}(obj, maxpixels = 1e+07, ...)

\method{as.imlist}{RasterStackBrick}(obj, maxpixels = 1e+07, ...)
}
\arguments{
\item{obj}{an object of class "RasterLayer"}

\item{maxpixels}{max. number of pixels to load  (default 1e7)}

\item{...}{ignored}
}
\description{
The raster library stores its data as "RasterLayer" and "RasterBrick" objects. The raster package can store its data  out-of-RAM, so in order not to load too much data the "maxpixels" argument sets a limit on how many pixels are loaded.
}
\author{
Simon Barthelme, adapted from the image method for RasterLayer by Robert J Hijmans
}
