% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imageR.R
\name{renderImageR}
\alias{renderImageR}
\title{Widget render function for use in Shiny}
\usage{
renderImageR(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

