% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renorm}
\alias{renorm}
\title{Renormalise image}
\usage{
renorm(x, min = 0, max = 255)
}
\arguments{
\item{x}{numeric data}

\item{min}{min of the range}

\item{max}{max of the range}
}
\description{
Pixel data is usually expressed on a 0...255 scale for displaying. This function performs a linear renormalisation to range min...max
}
\examples{
renorm(0:10)
renorm(-5:5) #Same as above
}
\author{
Simon Barthelme
}
