% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resize_doubleXY}
\alias{resize_doubleXY}
\title{Resize image to double-size, using the Scale2X algorithm.}
\usage{
resize_doubleXY(im)
}
\description{
Use anisotropic upscaling algorithm
      <a href="http://scale2x.sourceforge.net/algorithm.html">described here</a>.
}

