% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resize_tripleXY}
\alias{resize_tripleXY}
\title{Resize image to triple-size, using the Scale2X algorithm.}
\usage{
resize_tripleXY(im)
}
\description{
Use anisotropic upscaling algorithm
      <a href="http://scale2x.sourceforge.net/algorithm.html">described here</a>.
}

