% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotate}
\alias{rotate}
\title{Rotate image by an arbitrary angle.}
\usage{
rotate(im, angle, interpolation = 1L, boundary = 0L)
}
\arguments{
\item{angle}{Rotation angle, in degrees.}

\item{interpolation}{Type of interpolation. Can be <tt>{ 0=nearest | 1=linear | 2=cubic }</tt>.}

\item{boundary}{Boundary conditions. Can be <tt>{  0=dirichlet | 1=neumann | 2=periodic }</tt>.}
}
\description{
Most of the time, the size of the image is modified.
}

