% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{select_patches}
\alias{select_patches}
\title{Return image patches centered at cx,cy with width wx and height wy}
\usage{
select_patches(im, cx, cy, wx, wy)
}
\arguments{
\item{cx,cy:}{vector of coordinates for patch centers}

\item{wx,wy:}{vector of coordinates for patch width and height}
}
\description{
Return image patches centered at cx,cy with width wx and height wy
}

