% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{split_connected}
\alias{split_connected}
\title{Split pixset into connected components}
\usage{
split_connected(px, ...)
}
\arguments{
\item{px}{a pixset}

\item{...}{further arguments passed to label}
}
\value{
a list of pixsets
}
\description{
Compute connected components (using "label"), then split into as many sets as there are components.
Useful for segmentation
}
\examples{
px <- isoblur(grayscale(boats),5) > .75
plot(px)
spl <- split_connected(px)
plot(spl[[1]])
px <- isoblur(grayscale(boats),5) > .75
plot(px)
spl <- split_connected(px)
plot(spl[[1]])
}
\seealso{
label
}
\author{
Simon Barthelme
}
