% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{stencil.cross}
\alias{stencil.cross}
\title{A cross-shaped stencil}
\usage{
stencil.cross(z = FALSE, cc = FALSE, origin = FALSE)
}
\arguments{
\item{z}{include neighbours along the z axis}

\item{cc}{include neighbours along the cc axis}

\item{origin}{include center pixel (default false)}
}
\value{
a data.frame defining a stencil
}
\description{
Returns a stencil corresponding to all nearest-neighbours of a pixel
}
\author{
Simon Barthelme
}
\seealso{
get.stencil
}

