% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{[.cimg}
\alias{[.cimg}
\title{Array subset operator for cimg objects}
\usage{
\method{[}{cimg}(x, ...)
}
\arguments{
\item{x}{an image (cimg object)}

\item{...}{subsetting arguments}
}
\description{
Works mostly just like the regular array version of x[...], the only difference being that it returns cimg objects when it makes sense to do so. For example im[,,,1] is just like as.array(im)[,,,1] except it returns a cimg object (containing only the first colour channel)
}
\seealso{
imsub, which provides a more convenient interface, crop
}

