% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixset.R
\name{where}
\alias{where}
\title{Return locations in pixel set}
\usage{
where(x)
}
\arguments{
\item{x}{a pixset}
}
\description{
Return locations in pixel set
}
\examples{
#All pixel locations with value greater than .99
where(boats > .99) 
}
