% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{add.colour}
\alias{add.colour}
\alias{add.color}
\title{Add colour channels to a grayscale image or pixel set}
\usage{
add.colour(im, simple = TRUE)

add.color(im, simple = TRUE)
}
\arguments{
\item{im}{a grayscale image}

\item{simple}{if TRUE just stack three copies of the grayscale image, if FALSE treat the image as the L channel in an HSL representation. Default TRUE. For pixel sets this option makes no sense and is ignored.}
}
\value{
an image of class cimg
}
\description{
Add colour channels to a grayscale image or pixel set
}
\section{Functions}{
\itemize{
\item \code{add.color}: Alias for add.colour
}}

\examples{
grayscale(boats) #No more colour channels
add.colour(grayscale(boats)) #Image has depth = 3 (but contains only grays)
}
\author{
Simon Barthelme
}
