% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{add.colour}
\alias{add.colour}
\title{Add colour channels to an grayscale image}
\usage{
add.colour(im)
}
\arguments{
\item{im}{a grayscale image}
}
\value{
an image of class cimg
}
\description{
Add colour channels to an grayscale image
}
\author{
Simon Barthelme
}

