% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{as.cimg.array}
\alias{as.cimg.array}
\title{Turn an numeric array into a cimg object}
\usage{
\method{as.cimg}{array}(X)
}
\arguments{
\item{X}{an array}
}
\description{
If the array has two dimensions, we assume it's a grayscale image. If it has three dimensions we assume it's a video, unless the third dimension has a depth of 3, in which case we assume it's a colour image,
}

