% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{as.cimg.data.frame}
\alias{as.cimg.data.frame}
\title{Create an image from a data.frame}
\usage{
\method{as.cimg}{data.frame}(df, v.name = "value", dims)
}
\arguments{
\item{df}{a data.frame}

\item{v.name}{name of the variable to extract pixel values from (default "value")}

\item{dims}{a vector of length 4 corresponding to image dimensions. If missing, a guess will be made.}
}
\value{
an object of class cimg
}
\description{
The data frame must be of the form (x,y,value) or (x,y,z,value), or (x,y,z,cc,value). The coordinates must be valid image coordinates (i.e., positive integers).
}
\author{
Simon Barthelme
}

