% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{as.data.frame.cimg}
\alias{as.data.frame.cimg}
\title{Convert a pixel image to a data.frame}
\usage{
\method{as.data.frame}{cimg}(im)
}
\arguments{
\item{im}{an image of class cimg}
}
\value{
a data.frame
}
\description{
This function combines the output of pixel.grid with the actual values (stored in $value)
}
\author{
Simon Barthelme
}

