% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autocrop}
\alias{autocrop}
\title{Autocrop image region}
\usage{
autocrop(im, color = c(0, 0, 0), axes = "zyx")
}
\arguments{
\item{im}{an image}

\item{color}{Color used for the crop. If  0, color is guessed.}

\item{axes}{Axes used for the crop.}
}
\description{
Autocrop image region
}
\examples{
#Add pointless padding
padded <- pad(boats,30,"xy")
plot(padded)
#Remove padding
autocrop(padded,color=c(0,0,0)) \%>\% plot
}

