% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{autocrop}
\alias{autocrop}
\title{Autocrop image region}
\usage{
autocrop(im, color, axes = "zyx")
}
\arguments{
\item{color}{Color used for the crop. If  0, color is guessed.}

\item{axes}{Axes used for the crop.}
}
\description{
Autocrop image region
}

