% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boxblur}
\alias{boxblur}
\title{Blur image with a box filter (square window)}
\usage{
boxblur(im, boxsize, neumann = TRUE)
}
\arguments{
\item{im}{an image}

\item{boxsize}{Size of the box window (can be subpixel).}

\item{neumann}{If true, use Neumann boundary conditions, Dirichlet otherwise  (default true, Neumann)}
}
\description{
Blur image with a box filter (square window)
}
\examples{
boxblur(boats,5) \%>\% plot(main="Dirichlet boundary")
boxblur(boats,5,TRUE) \%>\% plot(main="Neumann boundary")
}
\seealso{
deriche(), vanvliet().
}
