% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boxblur_xy}
\alias{boxblur_xy}
\title{Blur image with a box filter.}
\usage{
boxblur_xy(inp, sx, sy, boundary_conditions = TRUE)
}
\arguments{
\item{boundary_conditions}{Boundary conditions. Can be <tt>{ false=dirichlet | true=neumann }</tt>.}

\item{sigma_x}{Size of the box window, along the X-axis.}

\item{sigma_y}{Size of the box window, along the Y-axis.}

\item{sigma_z}{Size of the box window, along the Z-axis.}
}
\description{
This is a recursive algorithm, not depending on the values of the box kernel size.
}
\seealso{
blur().
}

