% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boxblur_xy}
\alias{boxblur_xy}
\title{Blur image with a box filter.}
\usage{
boxblur_xy(im, sx, sy, neumann = TRUE)
}
\arguments{
\item{im}{an image}

\item{sx}{Size of the box window, along the X-axis.}

\item{sy}{Size of the box window, along the Y-axis.}

\item{neumann}{If true, use Neumann boundary conditions, Dirichlet otherwise  (default true, Neumann)}
}
\description{
This is a recursive algorithm, not depending on the values of the box kernel size.
}
\examples{
boxblur_xy(boats,20,5) \%>\% plot(main="Anisotropic blur")
}
\seealso{
blur().
}
