% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{channels}
\alias{channels}
\title{Split a colour image into a list of separate channels}
\usage{
channels(im, index, drop = FALSE)
}
\arguments{
\item{im}{an image}

\item{index}{which channels to extract (default all)}

\item{drop}{if TRUE drop extra dimensions, returning normal arrays and not cimg objects}
}
\value{
a list of channels
}
\description{
Split a colour image into a list of separate channels
}
\seealso{
frames
}

