% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{cimg.extract}
\alias{B}
\alias{G}
\alias{R}
\alias{channel}
\alias{cimg.extract}
\alias{frame}
\title{Various shortcuts for extracting colour channels, frames, etc}
\usage{
frame(im, index)

channel(im, ind)

R(im)

G(im)

B(im)
}
\arguments{
\item{im}{an image}

\item{index}{frame index}

\item{ind}{channel index}
}
\description{
Various shortcuts for extracting colour channels, frames, etc

Extract one frame out of a 4D image/video
}
\section{Functions}{
\itemize{
\item \code{frame}: 

\item \code{channel}: Extract an image channel

\item \code{R}: Extract red channel

\item \code{G}: Extract green channel

\item \code{B}: Extract blue channel
}}
\author{
Simon Barthelme
}

