% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cimg.use.openmp}
\alias{cimg.use.openmp}
\title{Control CImg's parallelisation}
\usage{
cimg.use.openmp(mode = "adaptive")
}
\arguments{
\item{mode}{Either "adaptive","always" or "none". The default is adaptive (parallelisation for large images only).}
}
\value{
NULL (function is used for side effects)
}
\description{
On supported architectures CImg can parallelise many operations using OpenMP.
Use this function to turn parallelisation on or off.
}
\examples{
cimg.use.openmp("never") #turn off parallelisation
}
\author{
Simon Barthelme
}
