% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve}
\alias{convolve}
\title{Convolve image by a mask.}
\usage{
convolve(im, filter, boundary_conditions = TRUE, is_normalised = FALSE)
}
\arguments{
\item{im}{an image}

\item{boundary_conditions}{= the border condition type (0=zero, 1=dirichlet)}

\item{mask}{= the convolution kernel.}

\item{is_normalized}{= enable local normalization.}
}
\description{
The result  res of the convolution of an image img by a mask mask is defined to be:
      res(x,y,z) = sum_{i,j,k} img(x-i,y-j,z-k)*mask(i,j,k)
}

