% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correlate}
\alias{correlate}
\title{Correlate image by a mask.}
\usage{
correlate(im, filter, boundary_conditions = TRUE, is_normalised = FALSE)
}
\arguments{
\item{im}{an image}

\item{boundary_conditions}{= the border condition type (0=zero, 1=dirichlet)}

\item{mask}{= the correlation kernel.}

\item{is_normalized}{= enable local normalization.}
}
\description{
The correlation of the image instance this by the mask mask is defined to be:
 res(x,y,z) = sum_{i,j,k} (*this)(x + i,y + j,z + k)*mask(i,j,k).
}

