% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diffusion_tensors}
\alias{diffusion_tensors}
\title{Compute field of diffusion tensors for edge-preserving smoothing.}
\usage{
diffusion_tensors(im, sharpness = 0.7, anisotropy = 0.6, alpha = 0.6,
  sigma = 1.1, is_sqrt = FALSE)
}
\arguments{
\item{im}{an image}

\item{sharpness}{Sharpness}

\item{anisotropy}{Anisotropy}

\item{alpha}{Standard deviation of the gradient blur.}

\item{sigma}{Standard deviation of the structure tensor blur.}

\item{is_sqrt}{Tells if the square root of the tensor field is computed instead.}
}
\description{
Compute field of diffusion tensors for edge-preserving smoothing.
}

