% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dilate}
\alias{dilate}
\title{Dilate image by a structuring element.}
\usage{
dilate(im, mask, boundary_conditions = TRUE, is_normalised = FALSE)
}
\arguments{
\item{mask}{Structuring element.}

\item{boundary_conditions}{Boundary conditions.}

\item{is_normalized}{Sets if the erosion is locally normalized.}
}
\description{
Dilate image by a structuring element.
}

