% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{display.list}
\alias{display.list}
\title{Display image list using CImg library}
\usage{
\method{display}{list}(x, ...)
}
\arguments{
\item{x}{a list of cimg objects}

\item{...}{ignored}
}
\description{
Click on individual images to zoom in.
}
\examples{
##Not run: interactive only 
## imgradient(boats,"xy") \%>\% display
}
