% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{display}
\alias{display}
\title{Display object using CImg library}
\usage{
display(x, ...)
}
\arguments{
\item{x}{an image or a list of images}

\item{...}{ignored}
}
\description{
CImg has its own functions for fast, interactive image plotting. Use this if you get frustrated with slow rendering in RStudio.
}
\seealso{
display.cimg, display.imlist
}
