% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{display_list}
\alias{display_list}
\title{Display image list using CImg library}
\usage{
display_list(imlist)
}
\arguments{
\item{imlist}{a list of cimg objects}
}
\description{
Display image list using CImg library
}

