% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erode}
\alias{erode}
\title{Erode image by a structuring element.}
\usage{
erode(im, mask, boundary_conditions = TRUE, is_normalised = FALSE)
}
\arguments{
\item{mask}{Structuring element.}

\item{boundary_conditions}{Boundary conditions.}

\item{is_normalized}{Sets if the erosion is locally normalized.}
}
\description{
Erode image by a structuring element.
}

