% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grayscale}
\alias{grayscale}
\title{Convert an RGB image to grayscale}
\usage{
grayscale(im)
}
\value{
a grayscale image (spectrum == 1)
}
\description{
This function converts from RGB to grayscale by first converting to HSL and keeping only the L channel
}

