% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haar}
\alias{haar}
\title{Compute Haar multiscale wavelet transform.}
\usage{
haar(im, inverse = FALSE, nb_scales = 1L)
}
\arguments{
\item{im}{an image}

\item{nb_scales}{Number of scales used for the transform.}

\item{axis}{Axis considered for the transform.}

\item{invert}{Set inverse of direct transform.}
}
\description{
Compute Haar multiscale wavelet transform.
}

