% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{idply}
\alias{idply}
\title{Split an image along axis, apply function, return a data.frame}
\usage{
idply(im, axis, fun, ...)
}
\arguments{
\item{im}{image}

\item{axis}{axis for the split (e.g "c")}

\item{fun}{function to apply}

\item{...}{extra arguments to function fun}
}
\description{
Shorthand for imsplit followed by ldply
}
\examples{
parrots <- load.image(system.file('extdata/parrots.png',package='imager'))
idply(parrots,"c",mean) #mean luminance per colour channel
}

