% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{ilply}
\alias{ilply}
\title{Split an image along axis, apply function, return a list}
\usage{
ilply(im, axis, fun, ...)
}
\arguments{
\item{im}{image}

\item{axis}{axis for the split (e.g "c")}

\item{fun}{function to apply}

\item{...}{extra arguments for function fun}
}
\description{
Shorthand for imsplit followed by llply
}
\examples{
parrots <- load.image(system.file('extdata/parrots.png',package='imager'))
ilply(parrots,"c",mean) #mean luminance per colour channel
}

