% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{im2cimg}
\alias{im2cimg}
\title{Convert an image in spatstat format to an image in cimg format}
\usage{
im2cimg(img)
}
\arguments{
\item{img}{a spatstat image}
}
\value{
a cimg image
}
\description{
as.cimg.im is an alias for the same function
}
\author{
Simon Barthelme
}
