% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\docType{package}
\name{imager}
\alias{imager}
\alias{imager-package}
\title{imager: an R library for image processing, based on CImg}
\description{
CImg by David Tschumperlé is a C++ library for image processing. It provides most common functions for image manipulation and filtering, as well as some advanced algorithms. imager makes these functions accessible from R and adds some basic plotting and subsetting.
You should install ImageMagick if you want support for common image formats (png, jpg, etc.)
}

