% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{imdirac}
\alias{imdirac}
\title{Generates a "dirac" image, i.e. with all values set to 0 except one.}
\usage{
imdirac(dims, x, y, z = 1, cc = 1)
}
\arguments{
\item{dims}{a vector of image dimensions, or an image whose dimensions will be used}

\item{x}{where to put the dirac (x coordinate)}

\item{y}{y coordinate}

\item{z}{z coordinate (default 1)}

\item{cc}{colour coordinate (default 1)}
}
\value{
an image
}
\description{
This small utility is useful to examine the impulse response of a filter
}
\examples{
#Impulse response of the blur filter
imdirac(c(50,50,1,1),20,20) \%>\% isoblur(sigma=2)  \%>\% plot
#Impulse response of the first-order Deriche filter
imdirac(c(50,50,1,1),20,20) \%>\% deriche(sigma=2,order=1,axis="x")  \%>\% plot
}
\author{
Simon Barthelme
}

