% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imeval}
\alias{imeval}
\title{Pixel-wise evaluation of a CImg expression}
\usage{
imeval(inp, cmd)
}
\description{
Pixel-wise evaluation of a CImg expression
}
\examples{
imfill(10,10) \%>\% imeval('x+y') \%>\% plot
# Box filter
boxf = "v=0;for(iy=y-3,iy<y+3,iy++,for(ix=x-3,ix< x+3,ix++,v+=i(ix,iy)));v"
imeval(boats,boxf) \%>\% plot
# Example by D. Tschumperlé: Julia set
julia <-  "
   zr = -1.2 + 2.4*x/w;
   zi = -1.2 + 2.4*y/h;
   for (iter = 0, zr^2+zi^2<=4 && iter<256, iter++,
     t = zr^2 - zi^2 + 0.5;
     (zi *= 2*zr) += 0.2;
     zr = t
   );
   iter"
imfill(500,500) \%>\% imeval(julia) \%>\% plot
}

