% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{imgradient}
\alias{imgradient}
\title{Compute image gradient}
\usage{
imgradient(im, axes, scheme = 3)
}
\arguments{
\item{im}{an image of class cimg}

\item{scheme}{numerical scheme (default '3')}

\item{axes:}{direction along which to compute the gradient. Either a single character (e.g. "x"), or multiple characters (e.g. "xyz")}
}
\value{
an image or a list of images, depending on the value of "axes"
}
\description{
Light interface for get_gradient. Refer to get_gradient for details on the computation.
}
\author{
Simon Barthelmé
}

