% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imglist.R
\name{imlist}
\alias{as.imlist.list}
\alias{imlist}
\alias{is.imlist}
\title{Image list}
\usage{
imlist(l)

is.imlist(l)

as.imlist.list(l)
}
\arguments{
\item{l}{a list}

\item{...}{ignored}
}
\description{
An imlist object is simply a list of images (of class cimg). For convenience, some generic functions are defined that wouldn't work on plain lists, like plot, display and as.data.frame.
}
\examples{
#imsplit returns objects of class "imlist"
imsplit(boats,"c")
list(a=imfill(3,3),b=imfill(10,10)) \%>\% imlist
imsplit(boats,"x",6) \%>\% plot
}
\seealso{
plot.imlist, display.imlist, as.data.frame.imlist
}

