% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isoblur}
\alias{isoblur}
\title{Blur image isotropically.}
\usage{
isoblur(im, sigma, neumann = TRUE, gaussian = FALSE)
}
\arguments{
\item{im}{an image}

\item{sigma}{Standard deviation of the blur.}

\item{neumann}{If true, use Neumann boundary conditions, Dirichlet otherwise  (default true, Neumann)}

\item{gaussian}{Use a Gaussian filter (actually vanVliet-Young). Default: 0th-order Deriche filter.}
}
\description{
Blur image isotropically.
}
\examples{
isoblur(boats,3) \%>\% plot(main="Isotropic blur, sigma=3")
isoblur(boats,3) \%>\% plot(main="Isotropic blur, sigma=10")
}
\seealso{
deriche,vanvliet

medianblur
}

