% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{label}
\alias{label}
\title{Label connected components.}
\usage{
label(im, is_high_connectivity = FALSE, tolerance = 0)
}
\arguments{
\item{is_high_connectivity}{Boolean that choose between 4(false)- or 8(true)-connectivity
in 2d case, and between 6(false)- or 26(true)-connectivity in 3d case.}

\item{tolerance}{Tolerance used to determine if two neighboring pixels belong to the same region.}
}
\description{
The algorithm of connected components computation has been primarily done
by A. Meijster, according to the publication:
'W.H. Hesselink, A. Meijster, C. Bron, "Concurrent Determination of Connected Components.",
      In: Science of Computer Programming 41 (2001), pp. 173--194'.
}

