% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{load.image}
\alias{load.image}
\title{Load image from file or URL}
\usage{
load.image(file)
}
\arguments{
\item{file}{path to file}
}
\value{
an object of class 'cimg'
}
\description{
You'll need ImageMagick for formats other than PNG and JPEG. If the image is actually a video, you'll need ffmpeg. If the path is actually a URL, it should start with http(s) or ftp(s).
}
\examples{
#Find path to example file from package
fpath <- system.file('extdata/Leonardo_Birds.jpg',package='imager') 
im <- load.image(fpath) 
plot(im)
}

