% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{load.image}
\alias{load.image}
\title{Load image from file}
\usage{
load.image(file)
}
\arguments{
\item{file}{path to file}
}
\value{
an object of class 'cimg'
}
\description{
You'll need ImageMagick for formats other than PNG and JPEG. If the image is actually a video, you'll need ffmpeg.
}
\examples{
fpath <- system.file('extdata/Leonardo_Birds.jpg',package='imager') #Path to example file from package
im <- load.image(fpath)
plot(im)
}

