% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mclosing}
\alias{mclosing}
\title{Morphological closing (dilation followed by erosion)}
\usage{
mclosing(im, mask, boundary_conditions = TRUE, is_normalised = FALSE)
}
\arguments{
\item{mask}{Structuring element.}

\item{boundary_conditions}{Boundary conditions.}

\item{is_normalized}{Determines if the closing is locally normalized.}
}
\description{
Morphological closing (dilation followed by erosion)
}

