% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{medianblur}
\alias{medianblur}
\title{Blur image with the median filter.}
\usage{
medianblur(im, n, threshold)
}
\arguments{
\item{im}{an image}

\item{n}{Size of the median filter.}

\item{threshold}{Threshold used to discard pixels too far from the current pixel value in the median computation. Can be used for edge-preserving smoothing.}
}
\description{
Blur image with the median filter.
}
\examples{
medianblur(boats,5,Inf) \%>\% plot(main="Median blur, 5 pixels")
medianblur(boats,10,Inf) \%>\% plot(main="Median blur, 10 pixels")
medianblur(boats,10,8) \%>\% plot(main="Median blur, 10 pixels, threshold = 8")
}
\seealso{
isoblur, boxblur
}

