% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{medianblur}
\alias{medianblur}
\title{Blur image with the median filter.}
\usage{
medianblur(im, n, threshold)
}
\arguments{
\item{im}{an image}

\item{n}{Size of the median filter.}

\item{threshold}{Threshold used to discard pixels too far from the current pixel value in the median computation.}
}
\description{
Blur image with the median filter.
}

