% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mirror}
\alias{mirror}
\title{Mirror image content along specified axis}
\usage{
mirror(im, axis)
}
\arguments{
\item{axis}{Mirror axis ("x","y","z","c")}
}
\description{
Mirror image content along specified axis
}

