% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mopening}
\alias{mopening}
\title{Morphological opening (erosion followed by dilation)}
\usage{
mopening(im, mask, boundary_conditions = TRUE, is_normalised = FALSE)
}
\arguments{
\item{mask}{Structuring element.}

\item{boundary_conditions}{Boundary conditions.}

\item{is_normalized}{Determines if the opening is locally normalized.}
}
\description{
Morphological opening (erosion followed by dilation)
}

