% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{play}
\alias{play}
\title{Play a video}
\usage{
play(vid, loop = FALSE, delay = 30L)
}
\arguments{
\item{vid}{A cimg object, to be played as video}

\item{loop}{loop the video (default false)}

\item{delay}{delay between frames, in ms. Default 30.}
}
\description{
A very basic video player. Press the space bar to pause and ESC to close.
}

