% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{plot.cimg}
\alias{plot.cimg}
\title{Display an image using base graphics}
\usage{
\method{plot}{cimg}(im, frame, rescale.color = TRUE, ...)
}
\arguments{
\item{im}{the image}

\item{frame}{which frame to display, if the image has depth > 1}

\item{rescale.color}{rescale channels so that the values are in [0,1]}

\item{...}{other parameters to be passed to plot.default (eg "main")}
}
\description{
Display an image using base graphics
}
\seealso{
display, which is much faster
}

