% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resize_halfXY}
\alias{resize_halfXY}
\title{Resize image to half-size, using an optimized filter}
\usage{
resize_halfXY(im)
}
\description{
Use anisotropic upscaling algorithm
      <a href="http://scale2x.sourceforge.net/algorithm.html">described here</a>.
}

