% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm.alpha}
\alias{rm.alpha}
\title{Remove alpha channel and store as attribute}
\usage{
rm.alpha(im)
}
\arguments{
\item{im}{an image with 4 RGBA colour channels}
}
\value{
an image with only three RGB channels and the alpha channel as attribute
}
\description{
Remove alpha channel and store as attribute
}
\examples{
#An image with 4 colour channels (RGBA)
im <- imfill(2,2,val=c(0,0,0,0))
#Remove fourth channel
rm.alpha(im)
attr(rm.alpha(im),"alpha")
}
\seealso{
flatten.alpha
}
\author{
Simon Barthelme
}
