% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{save.image}
\alias{save.image}
\title{Save image}
\usage{
save.image(im, file, quality = 0.7)
}
\arguments{
\item{im}{an image (of class cimg)}

\item{file}{path to file. The format is determined by the file's name}

\item{quality}{(JPEG only) default 0.7. Higher quality means less compression.}
}
\value{
nothing
}
\description{
You'll need ImageMagick for formats other than PNG and JPEG.
}
\examples{
#Create temporary file
tmpF <- tempfile(fileext=".png")
#Save boats image
save.image(boats,tmpF)
#Read back and display
load.image(tmpF) \%>\% plot
}
\seealso{
save.video
}
