% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{save.image}
\alias{save.image}
\title{Save image}
\usage{
save.image(im, file)
}
\arguments{
\item{im}{an image (of class cimg)}

\item{file}{path to file. The format is determined by the file's name}
}
\value{
nothing
}
\description{
You'll need ImageMagick for formats other than PNG and JPEG. If the image is actually a video, you'll need ffmpeg.
}

