% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cimg_class.R
\name{squeeze}
\alias{squeeze}
\title{Remove empty dimensions from an array}
\usage{
squeeze(x)
}
\arguments{
\item{x}{an array}
}
\description{
Works just like Matlab's squeeze function: if anything in dim(x) equals one the corresponding dimension is removed
}

