% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{watershed}
\alias{watershed}
\title{Compute watershed transform.}
\usage{
watershed(im, priority, fill_lines = TRUE)
}
\arguments{
\item{priority}{Priority map.}

\item{fill_lines}{Sets if watershed lines must be filled or not.}
}
\description{
Non-zero values are propagated to zero-valued ones according to
      the priority map.
}

